java-spring: documentManager interface separation
