+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/util/documentManagers/DocumentTokenManagerImpl.java
@@ -1,0 +1,44 @@
+package com.onlyoffice.integration.util.documentManagers;
+
+import org.primeframework.jwt.Signer;
+import org.primeframework.jwt.Verifier;
+import org.primeframework.jwt.domain.JWT;
+import org.primeframework.jwt.hmac.HMACSigner;
+import org.primeframework.jwt.hmac.HMACVerifier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.util.Map;
+
+@Component
+public class DocumentTokenManagerImpl implements DocumentTokenManager {
+
+    @Value("${files.docservice.secret}")
+    private String tokenSecret;
+
+    public String createToken(Map<String, Object> payloadClaims) {
+        try {
+            Signer signer = HMACSigner.newSHA256Signer(tokenSecret);
+            JWT jwt = new JWT();
+            for (String key : payloadClaims.keySet()) {
+                jwt.addClaim(key, payloadClaims.get(key));
+            }
+            return JWT.getEncoder().encode(jwt, signer);
+        } catch (Exception e) {
+            return "";
+        }
+    }
+
+    public boolean tokenEnabled() {
+        return tokenSecret != null && !tokenSecret.isEmpty();
+    }
+
+    public JWT readToken(String token) {
+        try {
+            Verifier verifier = HMACVerifier.newVerifier(tokenSecret);
+            return JWT.getDecoder().decode(token, verifier);
+        } catch (Exception exception) {
+            return null;
+        }
+    }
+}
